LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-name-shadow-expected.txt


FAIL Outer animation can see scroll timeline defined by :host assert_equals: expected (string) "y" but got (undefined) undefined
FAIL Outer animation can see scroll timeline defined by ::slotted assert_equals: expected (string) "y" but got (undefined) undefined
FAIL Inner animation can see scroll timeline defined by ::part assert_equals: expected (string) "x" but got (undefined) undefined
FAIL Slotted element can see scroll timeline within the shadow assert_equals: expected (string) "y" but got (undefined) undefined
